# node/attendanceMgm/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TML app",
    "description": "APIs for TML app",
    "version": "1.0"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json",
    "text/xml"
  ],
  "host": "localhost:8080",
  "basePath": "/v1/tml",
  "paths": {    
...
    "/users/{id}/attendance": {
      "get": {
        "x-swagger-router-controller": "UserController",
        "operationId": "getUserAttendance",
        "tags": [
          "/user"
        ],
        "description": "Returns specific user's attendance.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id",
            "required": true,
            "type": "string"
          },
          {
            "name": "startdate",
            "in": "query",
            "description": "startdate",
            "required": false,
            "type": "string"
          },
          {
            "name": "enddate",
            "in": "query",
            "description": "enddate",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/AttendanceList"
            }
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
	  "options": {
        "x-swagger-router-controller": "UserController",
        "operationId": "handleOption",
        "tags": [
          "/user"
        ],
        "description": "add new SOS",
		"parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id",
            "required": true,
            "type": "string"
          },
          {
            "name": "startdate",
            "in": "query",
            "description": "startdate",
            "required": false,
            "type": "string"
          },
          {
            "name": "enddate",
            "in": "query",
            "description": "enddate",
            "required": false,
            "type": "string"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    } ,
	 "/users/{user_id}": {
      "put": {
        "x-swagger-router-controller": "UserController",
        "operationId": "updateAlertStatus",
        "tags": [
          "/user"
        ],
        "description": "update alert status",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "user id",
            "required": true,
            "type": "string"
          },          
		{
            "name": "userObject",
            "description": "The new Entity to create",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User_PUT"
            }
          }		  
        ],
        "responses": {
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
	  "get": {
        "x-swagger-router-controller": "UserController",
        "operationId": "getAlertDetails",
        "tags": [
          "/user"
        ],
        "description": "Returns specific user's alert details",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "user id",
            "required": true,
            "type": "string"
          }          
        ],
        "responses": {
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
	  "options": {
        "x-swagger-router-controller": "UserController",
        "operationId": "handleOption",
        "tags": [
          "/user"
        ],
        "description": "Update alert status",
		"parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "user id",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "description": "status",
            "required": false,
            "type": "string"
          }          
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
     "/users": {      
      "post": {
        "x-swagger-router-controller": "UserController",
        "operationId": "addUser",
        "tags": [
          "/user"
        ],
        "description": "add new User",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userObject",
            "description": "The new Entity to create",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
	  "options": {
        "x-swagger-router-controller": "UserController",
        "operationId": "handleOption",
        "tags": [
          "/user"
        ],
        "description": "add new user details",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },    
...
  },
  "definitions": {
    "Error": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": [
        "message"
      ]
    },    
...
    "AttendanceList": {
      "properties": {
        "result": {
          "type": "string"
        },
        "JSON": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Attendance"
          }
        }
      },
      "required": [
        "result"
      ]
    },
    "Attendance": {
      "properties": {
        "date": {
          "type": "string"
        },
        "parking_slot": {
          "type": "integer"
        },
        "bin_c": {
          "type": "integer"
        },
        "user_id": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }, 
        "event": {
          "type": "string"
        },
        "fence": {
          "type": "string"
        },    
        "created_by": {
          "type": "string"
        },        
        "updated_by": {
          "type": "string"
        }
      }
    },
...
    "User": {
      "properties": {
        "full_name": {
          "type": "string"
        },
        "using_car": {
          "type": "boolean"
        },
        "user_id": {
          "type": "string"
        },
        "mobileno": {
          "type": "integer"
        }		
      }
    },   
...
	"User_PUT": {
      "properties": {
        "status": {
          "type": "string"
        }        
      }
    },	
...
  }
}